refactor:  InnerCurrentLocation, InnerSearch 분리
